Add multiline font-lock for function names
This is implemented in two parts:

Firstly the font-lock region needs to be extended where a function
name is at the beginning or a function keyword is at the end.

Secondly the font-lock matcher for a function keyword includes an
anchored match for the function name, where the search boundary is
potentially extended across multiple lines.