Dynamically classify syntax type for "?"
Classify the "?" character as a syntax character by default.

Use a syntax propertize function to reclassify the "?" character as
punctuation where it does not define its own symbol boundary (i.e. it
is adjacent to a symbol character which is not "?" and distorting the
symbol name).

Use a custom `thing-at-point' replacement which checks for "?" as a
self-delimiting symbol character (the "?" command which outputs CRLF).