Build with Autotools and add manual
Convert the package commentary to a Texinfo manual.

The README file is now generated from smaller portions of the manual,
rather than being an export the package commentary.

The commentary link in the customization group is replaced with a link
to the Info manual.

Convert the package news to a ChangeLog file, formatted to work with
outline mode.