Extend the Outline match to include the heading name
This allows the highlighting available by configuring
outline-minor-mode-highlight to include what is effectively the
heading name, rather than highlight only the syntax which indicated
that a heading is present.

Note that configuring outline-minor-mode-highlight to append Outline
Mode's font-lock faces appears to introduce significant slowdown.
Choosing to override the face entirely does not appear to suffer the
same issue.